requirements.txt
tensorflow-gpu==1.1.0
pandas
numpy

